# swagger.yaml
...
swagger: "2.0"
info:
  description: |
    This is a swagger defenition for [Telegram bot API](https://core.telegram.org/bots/api).
  version: "3.4"
  title: Telegram bot API
...
host: api.telegram.org
schemes:
 - https
consumes:
- application/json
produces:
- application/json
basePath: /
parameters:
  token:
    in: path
    name: token
    required: false
    type: string
    description: bot's token to authorize the request
paths:
  /bot{token}/getUpdates:
    get:
      tags:
      - updates
      description: ""
      operationId: getUpdates
      parameters:
      - $ref: '#/parameters/token'
      - in: query
        name: offset
        type: integer
        required: false
      - in: query
        name: limit
        type: integer
        required: false
      - in: query
        name: timeout
        type: integer
        required: false
      - in: query
        name: allowed_updates
        required: false
        type: array
        collectionFormat: multi
        items:
          # go-swagger bug here
          # cannot use $ref: "#/definitions/AllowedUpdate"
          type: string
          enum:
          - message
          - edited_message
          - channel_post
          - edited_channel_post
          - inline_query
          - chosen_inline_result
          - callback_query
      responses:
        200:
          description: An array of Updates
          schema:
            $ref: "#/definitions/ResponseUpdate"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        401:
          description: Unauthorized
          schema:
            $ref: "#/definitions/Error"
        403:
          description: Forbidden
          schema:
            $ref: "#/definitions/Error"
        404:
          description: Not Found
          schema:
            $ref: "#/definitions/Error"
        420:
          description: Flood
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal
          schema:
            $ref: "#/definitions/Error"
...
  /bot{token}/deleteChatPhoto:
    post:
      tags:
      - chats
      description: ""
      operationId: deleteChatPhoto
      parameters:
      - $ref: '#/parameters/token'
      - in: query
        name: chat_id
        type: string
        required: true
        description: ""
      responses:
        200:
          description: ""
          schema:
            $ref: "#/definitions/ResponseBool"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/Error"
        401:
          description: Unauthorized
          schema:
            $ref: "#/definitions/Error"
        403:
          description: Forbidden
          schema:
            $ref: "#/definitions/Error"
        404:
          description: Not Found
          schema:
            $ref: "#/definitions/Error"
        420:
          description: Flood
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal
          schema:
            $ref: "#/definitions/Error"
...
definitions:
  User:
    type: object
    properties:
      id:
        type: integer
      username:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      language_code:
        type: string
      is_bot:
        type: boolean
  Chat:
    type: object
    required:
    - id
    - type
    properties:
      id:
        type: integer
        x-nullable: false
      type:
        type: string
        enum:
        - private
        - group
        - supergroup
        - channel
      title:
        type: string
      username:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      all_members_are_administrators:
        type: boolean
      photo:
        $ref: '#/definitions/ChatPhoto'
      description:
        type: string
      invite_link:
        type: string
      pinned_message:
        $ref: '#/definitions/Message'
      sticker_set_name:
        type: string
      can_set_sticker_set:
        type: boolean
  Message:
    type: object
    properties:
      message_id:
        type: integer
      from:
        $ref: '#/definitions/User'
      date:
        type: integer
      chat:
        $ref: '#/definitions/Chat'
      forward_from:
        $ref: '#/definitions/User'
      forward_from_chat:
        $ref: '#/definitions/Chat'
      forward_from_message_id:
        type: integer
      forward_date:
        type: integer
      reply_to_message:
        $ref: '#/definitions/Message'
      edit_date:
        type: integer
      text:
        type: string
      entities:
        type: array
        items:
          $ref: '#/definitions/MessageEntity'
      caption_entities:
        type: array
        items:
          $ref: '#/definitions/MessageEntity'
      audio:
        $ref: '#/definitions/Audio'
      document:
        $ref: '#/definitions/Document'
      game:
        $ref: '#/definitions/Game'
      photo:
        type: array
        items:
          $ref: '#/definitions/PhotoSize'
      sticker:
        $ref: '#/definitions/Sticker'
      video:
        $ref: '#/definitions/Video'
      voice:
        $ref: '#/definitions/Voice'
      video_note:
        $ref: '#/definitions/VideoNote'
      caption:
        type: string
      contact:
        $ref: '#/definitions/Contact'
      location:
        $ref: '#/definitions/Location'
      venue:
        $ref: '#/definitions/Venue'
      new_chat_members:
        type: array
        items:
          $ref: '#/definitions/User'
      left_chat_member:
        $ref: '#/definitions/User'
      new_chat_title:
        type: string
      new_chat_photo:
        type: array
        items:
          $ref: '#/definitions/PhotoSize'
      delete_chat_photo:
        type: boolean
      group_chat_created:
        type: boolean
      supergroup_chat_created:
        type: boolean
      channel_chat_created:
        type: boolean
      migrate_to_chat_id:
        type: integer
      migrate_from_chat_id:
        type: integer
      pinned_message:
        $ref: '#/definitions/Message'
      invoice:
        $ref: '#/definitions/Invoice'
      successful_payment:
        $ref: '#/definitions/SuccessfulPayment'
      forward_signature:
        type: string
      author_signature:
        type: string
  MessageEntity:
    type: object
    properties:
      type:
        type: string
        enum:
        - mention
        - hashtag
        - bot_command
        - url
        - email
        - bold
        - italic
        - code
        - pre
        - text_link
        - text_mention
      offset:
        type: integer
      length:
        type: integer
      url:
        type: string
      user:
        $ref: '#/definitions/User'
  PhotoSize:
    type: object
    properties:
      file_id:
        type: string
      width:
        type: integer
      height:
        type: integer
      file_size:
        type: integer
  Audio:
    type: object
    properties:
      file_id:
        type: string
      duration:
        type: integer
      performer:
        type: string
      title:
        type: string
      mime_type:
        type: string
      file_size:
        type: integer
  Document:
    type: object
    properties:
      file_id:
        type: string
      thumb:
        $ref: '#/definitions/PhotoSize'
      file_name:
        type: string
      mime_type:
        type: string
      file_size:
        type: integer
  Sticker:
    type: object
    properties:
      file_id:
        type: string
      width:
        type: integer
      height:
        type: integer
      thumb:
        $ref: '#/definitions/PhotoSize'
      emoji:
        type: string
      set_name:
        type: string
      mask_position:
        $ref: '#/definitions/MaskPosition'
      file_size:
        type: integer
  Video:
    type: object
    properties:
      file_id:
        type: string
      width:
        type: integer
      height:
        type: integer
      duration:
        type: integer
      thumb:
        $ref: '#/definitions/PhotoSize'
      mime_type:
        type: string
      file_size:
        type: integer
  Voice:
    type: object
    properties:
      file_id:
        type: string
      duration:
        type: integer
      mime_type:
        type: string
      file_size:
        type: integer
  Contact:
    type: object
    properties:
      phone_number:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      user_id:
        type: integer
  Location:
    type: object
    properties:
      longitude:
        type: number
      latitude:
        type: number
  Venue:
    type: object
    properties:
      location:
        $ref: '#/definitions/Location'
      title:
        type: string
      address:
        type: string
      foursquare_id:
        type: string
...
  CallbackQuery:
    type: object
    properties:
      id:
        type: string
      from:
        $ref: '#/definitions/User'
      message:
        $ref: '#/definitions/Message'
      inline_message_id:
        type: string
      chat_instance:
        type: string
      data:
        type: string
      game_short_name:
        type: string
...
  Update:
    type: object
    properties:
      update_id:
        type: integer
      message:
        $ref: '#/definitions/Message'
      edited_message:
        $ref: '#/definitions/Message'
      channel_post:
        $ref: '#/definitions/Message'
      edited_channel_post:
        $ref: '#/definitions/Message'
      inline_query:
        $ref: '#/definitions/InlineQuery'
      chosen_inline_result:
        $ref: '#/definitions/ChosenInlineResult'
      callback_query:
        $ref: '#/definitions/CallbackQuery'
      shipping_query:
        $ref: '#/definitions/ShippingQuery'
      pre_checkout_query:
        $ref: '#/definitions/PreCheckoutQuery'
...
  ChosenInlineResult:
    type: object
    required:
    - result_id
    - from
    - query
    properties:
      result_id:
        type: string
      from:
        $ref: '#/definitions/User'
      location:
        $ref: '#/definitions/Location'
      inline_message_id:
        type: string
      query:
        type: string
  Game:
    type: object
    required:
    - title
    - description
    - photo
    properties:
      title:
        type: string
      description:
        type: string
      photo:
        type: array
        items:
          $ref: '#/definitions/PhotoSize'
      text:
        type: string
      text_entities:
        type: array
        items:
          $ref: '#/definitions/MessageEntity'
      animation:
        $ref: '#/definitions/Animation'
  Animation:
    type: object
    required:
    - file_id
    properties:
      file_id:
        type: string
      thumb:
        $ref: '#/definitions/PhotoSize'
      file_name:
        type: string
      mime_type:
        type: string
      file_size:
        type: integer
...
  InlineQuery:
    type: object
    properties:
      id:
        type: string
      from:
        $ref: '#/definitions/User'
      location:
        $ref: '#/definitions/Location'
      query:
        type: string
      offset:
        type: string
...
  ResponseBool:
    type: object
    properties:
      result:
        type: boolean
      error_code:
        type: integer
      ok:
        type: boolean
      description:
        type: string
  ResponseUpdate:
    type: object
    properties:
      result:
        type: array
        items:
          $ref: "#/definitions/Update"
      error_code:
        type: integer
      ok:
        type: boolean
      description:
        type: string
  Error:
    type: object
    properties:
      error_code:
        type: integer
      ok:
        type: boolean
      description:
        type: string
...
  AllowedUpdate:
    type: string
    enum:
    - message
    - edited_message
    - channel_post
    - edited_channel_post
    - inline_query
    - chosen_inline_result
    - callback_query
...
  VideoNote:
    type: object
    properties:
      file_id:
        type: string
      length:
        type: integer
      duration:
        type: integer
      thumb:
        $ref: '#/definitions/PhotoSize'
      file_size:
        type: integer
...
  Invoice:
    type: object
    properties:
      title:
        type: string
      description:
        type: string
      start_parameter:
        type: string
      currency:
        type: string
      total_amount:
        type: integer
  ShippingAddress:
    type: object
    properties:
      country_code:
        type: string
      stat:
        type: string
      city:
        type: string
      street_line1:
        type: string
      street_line2:
        type: string
      post_code:
        type: string
  OrderInfo:
    type: object
    properties:
      name:
        type: string
      phone_number:
        type: string
      email:
        type: string
      shipping_address:
        $ref: '#/definitions/ShippingAddress'
...
  SuccessfulPayment:
    type: object
    properties:
      currency:
        type: string
      total_amount:
        type: integer
      invoice_payload:
        type: string
      shipping_option_id:
        type: string
      order_info:
        $ref: '#/definitions/OrderInfo'
      telegram_payment_charge_id:
        type: string
      provider_payment_charge_id:
        type: string
  ShippingQuery:
    type: object
    properties:
      id:
        type: string
      from:
        $ref: '#/definitions/User'
      invoice_payload:
        type: string
      shipping_address:
        $ref: '#/definitions/ShippingAddress'
  PreCheckoutQuery:
    type: object
    properties:
      id:
        type: string
      from:
        $ref: '#/definitions/User'
      currency:
        type: string
      total_amount:
        type: integer
      invoice_payload:
        type: string
      shipping_option_id:
        type: string
      order_info:
        $ref: '#/definitions/OrderInfo'
...
  ChatPhoto:
    type: object
    properties:
      small_file_id:
        type: string
      big_file_id:
        type: string
...
  MaskPosition:
    type: object
    properties:
      point:
        type: string
        enum:
        - forehead
        - eyes
        - mouth
        - chin
      x_shift:
        type: number
      y_shift:
        type: number
      zoom:
        type: number
...